Prevent SalInstance::Yield() from spinning the processor when the timer has not expired and their are no pending AWT events

+++ vcl/java/source/app/salinst.cxx
@@ -283,31 +283,35 @@ void SalInstance::Yield( BOOL bWait )
 	// Dispatch pending AWT events
 	if ( bWait && !ImplGetSVData()->maAppData.mbAppQuit )
 	{
-		ULONG nTimeout = 0;
-
-		// Determine timeout
-		if ( pSalData->mnTimerInterval )
+		for ( ; ; )
 		{
-			timeval aTimeout;
+			ULONG nTimeout = 0;
 
-			gettimeofday( &aTimeout, NULL );
-			if ( pSalData->maTimeout > aTimeout )
+			// Determine timeout
+			if ( pSalData->mnTimerInterval )
 			{
-				aTimeout = pSalData->maTimeout - aTimeout;
-				nTimeout = aTimeout.tv_sec * 1000 + aTimeout.tv_usec / 1000;
+				timeval aTimeout;
+	
+				gettimeofday( &aTimeout, NULL );
+				if ( pSalData->maTimeout > aTimeout )
+				{
+					aTimeout = pSalData->maTimeout - aTimeout;
+					nTimeout = aTimeout.tv_sec * 1000 + aTimeout.tv_usec / 1000;
+				}
+				if ( nTimeout < 10 )
+					nTimeout = 10;
 			}
-			if ( nTimeout < 10 )
-				nTimeout = 10;
-		}
 
-		while ( ( pEvent = pSalData->mpEventQueue->getNextCachedEvent( nTimeout, TRUE ) ) != NULL )
-		{
-			// Reset timeout
-			nTimeout = 0;
-
-			pEvent->dispatch();
-			com_sun_star_vcl_VCLGraphics::flushAll();
-			delete pEvent;
+			if ( ( pEvent = pSalData->mpEventQueue->getNextCachedEvent( nTimeout, TRUE ) ) != NULL )
+			{
+				pEvent->dispatch();
+				com_sun_star_vcl_VCLGraphics::flushAll();
+				delete pEvent;
+			}
+			else
+			{
+				break;
+			}
 		}
 
 		// Check timer
